services/mcp/requirements.txt
fastmcp
httpx
beautifulsoup4
google-api-python-client
google-auth-oauthlib
psycopg2-binary
pgvector
python-dotenv